Minor refactoring #mod-6
