+++ modulus.api/DoubleAlternativeMod10.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Modulus.Shared;
+using Modulus.Shared.Models;
 
 namespace Modulus.api
 {
