+++ modulus.api/ModulusCalcBase.cs
@@ -1,4 +1,6 @@
+using Modulus.api.Contracts;
 using Modulus.Shared;
+using Modulus.Shared.Models;
 
 namespace Modulus.api
 {
