+++ modulus.api/WeightTable.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Modulus.api.Exceptions;
-using Modulus.Shared;
+using Modulus.Shared.Enums;
+using Modulus.Shared.Models;
 
 namespace Modulus.api
 {
@@ -74,6 +75,5 @@ namespace Modulus.api
             return items;
         }
 
-      
-}
+    }
 }
