+++ modulus.web/modulus.web.csproj
@@ -2,6 +2,8 @@
 
 	<PropertyGroup>
 		<TargetFramework>netcoreapp2.2</TargetFramework>
+		<AssemblyName>Modulus.Web</AssemblyName>
+		<RootNamespace>Modulus.Web</RootNamespace>
 	</PropertyGroup>
 
 	<ItemGroup>
